Moved app-framework files into its own subfolder
