+++ source/app-framework/Fonts/IconsFontAwesome5.h
