+++ source/app-framework/Helpers.h
