+++ source/app-framework/Image.h
